requirements.txt
socket
tqdm
ftplib
request
decouple